Extracted common filters to own component
Filters extracted: category selector, year, chart type
Added counties.json file
