+++ src/app/components/Sidebar/administration_sidebar.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import {Link} from 'react-router';
 import {BackLink} from '../BackLink/index';
-import {RouterParams, mreportPath, ireportPath, creportPath} from '../../helpers/url_helper';
+import {RouteParams, mreportPath, ireportPath, creportPath} from '../../helpers/url_helper';
 
 const style = require('./style.css');
 
 interface Props {
-  params: RouterParams;
+  params: RouteParams;
   location?: any;
 }
