+++ src/app/components/Sidebar/counties.json
@@ -1,0 +1,166 @@
+[
+  {
+    "code": "AB",
+    "name": "Alba"
+  },
+  {
+    "code": "AR",
+    "name": "Arad"
+  },
+  {
+    "code": "AG",
+    "name": "Argeș"
+  },
+  {
+    "code": "BC",
+    "name": "Bacău"
+  },
+  {
+    "code": "BH",
+    "name": "Bihor"
+  },
+  {
+    "code": "BN",
+    "name": "Bistriţa-Năsăud"
+  },
+  {
+    "code": "BT",
+    "name": "Botoșani"
+  },
+  {
+    "code": "BV",
+    "name": "Brașov"
+  },
+  {
+    "code": "BR",
+    "name": "Brăila"
+  },
+  {
+    "code": "BZ",
+    "name": "Buzău"
+  },
+  {
+    "code": "CS",
+    "name": "Caraș-Severin"
+  },
+  {
+    "code": "CL",
+    "name": "Călărași"
+  },
+  {
+    "code": "CJ",
+    "name": "Cluj"
+  },
+  {
+    "code": "CT",
+    "name": "Constanţa"
+  },
+  {
+    "code": "CV",
+    "name": "Covasna"
+  },
+  {
+    "code": "DB",
+    "name": "Dâmboviţa"
+  },
+  {
+    "code": "DJ",
+    "name": "Dolj"
+  },
+  {
+    "code": "GL",
+    "name": "Galaţi"
+  },
+  {
+    "code": "GR",
+    "name": "Giurgiu"
+  },
+  {
+    "code": "GJ",
+    "name": "Gorj"
+  },
+  {
+    "code": "HR",
+    "name": "Harghita"
+  },
+  {
+    "code": "HD",
+    "name": "Hunedoara"
+  },
+  {
+    "code": "IL",
+    "name": "Ialomiţa"
+  },
+  {
+    "code": "IS",
+    "name": "Iași"
+  },
+  {
+    "code": "IF",
+    "name": "Ilfov"
+  },
+  {
+    "code": "MM",
+    "name": "Maramureş"
+  },
+  {
+    "code": "MH",
+    "name": "Mehedinţi"
+  },
+  {
+    "code": "MS",
+    "name": "Mureş"
+  },
+  {
+    "code": "NT",
+    "name": "Neamţ"
+  },
+  {
+    "code": "OT",
+    "name": "Olt"
+  },
+  {
+    "code": "PH",
+    "name": "Prahova"
+  },
+  {
+    "code": "SM",
+    "name": "Satu Mare"
+  },
+  {
+    "code": "SJ",
+    "name": "Sălaj"
+  },
+  {
+    "code": "SB",
+    "name": "Sibiu"
+  },
+  {
+    "code": "SV",
+    "name": "Suceava"
+  },
+  {
+    "code": "TR",
+    "name": "Teleorman"
+  },
+  {
+    "code": "TM",
+    "name": "Timiş"
+  },
+  {
+    "code": "TL",
+    "name": "Tulcea"
+  },
+  {
+    "code": "VS",
+    "name": "Vaslui"
+  },
+  {
+    "code": "VL",
+    "name": "Vâlcea"
+  },
+  {
+    "code": "VN",
+    "name": "Vrancea"
+  }
+]
