+++ src/app/components/Sidebar/style.css
@@ -47,5 +47,7 @@
 
   & ul.adminType {
     padding: 0;
+    height: 100%;
+    overflow-y: scroll;
   }
 }
