+++ src/app/containers/Home/index.tsx
@@ -2,21 +2,15 @@ import * as React from 'react';
 import {ContentHeader} from '../../components/ContentHeader/index';
 const style = require('./style.css');
 
-class Home extends React.Component<any, any> {
+export class Home extends React.Component<any, any> {
   public render() {
-    // const map: CountyColorMap = {
-    //   arad: 'red', bihor: 'white', alba: '#cfe1f6',
-    // };
     return (
       <div className={style.Home}>
         <ContentHeader title="Prezentare indicatori SNA"/>
-
-        Home
-        {/*<RomaniaMap width={500} colorMap={map} />*/}
-
+        <div className={style.main}>
+          <img src={require('./intro.png')} />
+        </div>
       </div>
     );
   }
 }
-
-export { Home }
