+++ src/app/containers/Home/style.css
@@ -1,8 +1,9 @@
 .Home {
   text-align: center;
 
-  & .title {
-    font-family: Roboto Slab;
-    font-size: 40px;
+  & .main {
+    padding: 40px 65px;
+    display: flex;
+    align-items: flex-start;
   }
 }
