+++ src/app/containers/Html/style.css
@@ -10,4 +10,5 @@ a:hover {
 
 .App {
   max-width: 1280px;
+  font-family: Roboto;
 }
