+++ src/app/containers/MinistryOverview/style.css
@@ -3,7 +3,6 @@
 
   & .main {
     padding: 40px 65px;
-    font-family: Roboto;
 
     & .title {
       font-size: 15px;
@@ -54,46 +53,6 @@
     & .content {
       display: flex;
       flex-direction: row;
-    }
-
-    & .category {
-      & :global(.dropdown) {
-        width: 100%;
-
-        & :global(.dropdown-toggle) {
-          display: flex;
-          justify-content: space-between;
-          align-items: center;
-        }
-
-        & .select_category {
-          font-size: 20px;
-          padding: 10px 10px 12px 16px;
-          width: 100%;
-          text-align: left;
-        }
-
-        & :global(.dropdown-menu) {
-          font-size: 20px;
-        }
-      }
-    }
-  }
-
-  & .button {
-    border: none;
-    cursor: pointer;
-    border-radius: 2px;
-    outline: none;
-    background-color: transparent;
-    padding: 5px 10px;
-
-    &.active {
-      background-color: #FFFFFF;
-    }
-
-    & svg {
-      pointer-events: none;
     }
   }
 }
