+++ src/app/routes/index.tsx
@@ -3,12 +3,12 @@ import { IndexRoute, Route } from 'react-router';
 import { App, Home, About, Help } from 'containers';
 import {AdministationSidebar} from '../components/Sidebar/administration_sidebar';
 import {MinistriesSidebar} from '../components/Sidebar/ministries_sidebar';
-import {SelectIntitution} from '../containers/SelectInstitution/index';
 import {Sidebar} from '../components/Sidebar/index';
 import {Report} from '../containers/Report/index';
 import {CountiesSidebar} from '../components/Sidebar/counties_sidebar';
 import {EmptySidebar} from '../components/Sidebar/empty';
 import {MinistryOverview} from '../containers/MinistryOverview/index';
+import {CountyOverview} from "../containers/CountyOverview/index";
 
 export default (
   <Route path="/" component={App}>
@@ -17,10 +17,7 @@ export default (
     <Route path="help" component={Help} />
     <Route path="report/:id/1(/:mid)" components={{main: MinistryOverview, sidebar: MinistriesSidebar}} />
     <Route path="report/:id/2(/:mid)" components={{main: Report, sidebar: EmptySidebar}} />
-    <Route path="report/:id/3(/:mid)" components={{main: Report, sidebar: CountiesSidebar}} />
-    <Route path="selectAdministration/:id/1" components={{main: Home, sidebar: MinistriesSidebar}} />
-    <Route path="selectAdministration/:id/2" component={Home} />
-    <Route path="selectAdministration/:id/3" component={Home} />
-    <Route path="selectAdministration/:id" components={{main: SelectIntitution, sidebar: AdministationSidebar}} />
+    <Route path="report/:id/3(/:mid)" components={{main: CountyOverview, sidebar: CountiesSidebar}} />
+    <Route path="selectAdministration/:id" components={{main: Home, sidebar: AdministationSidebar}} />
   </Route>
 );
